인터페이스 속도 반환 타입을 int에서 uint64_t로 변경하고, 유효성 검사를 추가하여 음수 값을 처리하도록 수정

+++ include/collectors/network_collector.h
@@ -18,7 +18,7 @@ private:
     string getIPAddress(int sock, const string &if_name);
     string getMACAddress(const string &if_name);
     string getInterfaceStatus(const string &if_name);
-    int getInterfaceSpeed(const string &if_name);
+    uint64_t getInterfaceSpeed(const string &if_name);
     int getInterfaceMTU(int sock, const string &if_name);
 
 public:
